refactor(experience): share the entry detail drawer, drop emoji from Plaza
Extract EntryDrawer (and the Badge/CreatorLine/freshness/retiredDaysLeft
helpers it needs) out of Plaza.tsx into components/ExperienceDetailDrawer.tsx,
so the chat page can open an entry without growing a second detail UI that
would drift from this one. Plaza.tsx: 526 -> 361 lines. The ?entry= deep link
is unchanged; it just goes through the shared component now.

Creator line is reworked: 「👤 发布 X · 🤖 来源 Y」 becomes 「来源：X；Y[AI]」.
The square AI badge is not decoration — agents carry ordinary human-looking
names ("审核", "回款"), so the name alone never says whether it is a person.

Empty state 📚 gives way to the same Tabler line icon the sidebar uses for
经验广场.
