+++ frontend/src/pages/Plaza.tsx
@@ -10,54 +10,10 @@ import React, { useMemo, useState, useEffect } from 'react';
 import { useSearchParams } from 'react-router-dom';
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
 import { useTranslation } from 'react-i18next';
+import { IconBuildingMonument } from '@tabler/icons-react';
 import { experienceApi, type ExperienceEntry } from '../services/api';
 import { DraftEditor, Drawer, EXP_FIELDS, secondaryBtn, type Draft } from '../components/ExperienceDraftEditor';
-
-const SCOPE_LABELS: Record<string, string> = { company: '全公司', department: '本部门', user: '指定人' };
-
-// 2026年7月9日; empty string for null/invalid.
-function fmtDate(s?: string | null): string {
-    if (!s) return '';
-    const d = new Date(s);
-    if (isNaN(d.getTime())) return '';
-    return `${d.getFullYear()}年${d.getMonth() + 1}月${d.getDate()}日`;
-}
-
-function freshness(entry: ExperienceEntry): { label: string; stale: boolean } {
-    if (entry.status !== 'published') return { label: '', stale: false };
-    if (!entry.last_reviewed_at) return { label: '未复核', stale: true };
-    const d = new Date(entry.last_reviewed_at);
-    const dateStr = `${d.getFullYear()}年${d.getMonth() + 1}月${d.getDate()}日`;
-    const age = Date.now() - d.getTime();
-    const stale = age > 90 * 86400000;
-    return { label: `${stale ? '复核超期' : '已复核'}（${dateStr}）`, stale };
-}
-
-const RETIRED_TTL_DAYS = 30;
-// Days left before a retired entry is auto-deleted (retired_at + 30d). null when not applicable.
-function retiredDaysLeft(entry: ExperienceEntry): number | null {
-    if (entry.status !== 'retired' || !entry.retired_at) return null;
-    const d = new Date(entry.retired_at);
-    if (isNaN(d.getTime())) return null;
-    const deadline = d.getTime() + RETIRED_TTL_DAYS * 86400000;
-    return Math.max(0, Math.ceil((deadline - Date.now()) / 86400000));
-}
-
-const badgeStyle = (bg: string, fg: string): React.CSSProperties => ({
-    display: 'inline-block', padding: '1px 7px', borderRadius: 10, fontSize: 11,
-    background: bg, color: fg, whiteSpace: 'nowrap',
-});
-
-function Badge({ children, tone = 'muted' }: { children: React.ReactNode; tone?: 'muted' | 'warn' | 'ok' | 'accent' }) {
-    const tones: Record<string, [string, string]> = {
-        muted: ['var(--bg-tertiary)', 'var(--text-secondary)'],
-        warn: ['var(--error-subtle)', 'var(--error)'],
-        ok: ['var(--success-subtle)', 'var(--success)'],
-        accent: ['var(--accent-subtle)', 'var(--accent-text)'],
-    };
-    const [bg, fg] = tones[tone];
-    return <span style={badgeStyle(bg, fg)}>{children}</span>;
-}
+import { EntryDrawer, Badge, CreatorLine, freshness, retiredDaysLeft, SCOPE_LABELS } from '../components/ExperienceDetailDrawer';
 
 const sicon = (d: string) => (
     <svg width="14" height="14" viewBox="0 0 16 16" fill="none" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round"><path d={d} /></svg>
@@ -272,7 +228,9 @@ function TeamView({ loading, entries, stats, trending, activeTag, onOpen, onNew,
                             </div>
                         ) : (
                             <div style={{ border: '1px dashed var(--border-default)', borderRadius: 'var(--radius-lg)', padding: '48px 24px', textAlign: 'center' }}>
-                                <div style={{ fontSize: 30, marginBottom: 10 }}>📚</div>
+                                <div style={{ display: 'flex', justifyContent: 'center', marginBottom: 10, color: 'var(--text-tertiary)' }}>
+                                    <IconBuildingMonument size={30} stroke={1.5} />
+                                </div>
                                 <div style={{ fontSize: 'var(--text-base)', color: 'var(--text-primary)', fontWeight: 600, marginBottom: 6 }}>
                                     {t('experience.emptyTitle', '还没有已发布的经验')}
                                 </div>
@@ -364,24 +322,6 @@ function MineView({ loading, cats, cat, setCat, entries, onOpen }: {
 }
 
 
-// PRD v3: every entry shows its two creators — the human who published it + the source agent.
-function CreatorLine({ entry }: { entry: ExperienceEntry }) {
-    const created = fmtDate(entry.created_at);
-    // Never-modified entries have a null updated_at → fall back to created, so 上次修改 always shows.
-    const updated = fmtDate(entry.updated_at) || created;
-    const hasCreators = !!(entry.created_by_name || entry.origin_agent_name);
-    return (
-        <div style={{ display: 'flex', alignItems: 'center', gap: 6, fontSize: 12, color: 'var(--text-tertiary)', flexWrap: 'wrap' }}>
-            {entry.created_by_name && <span>👤 发布 {entry.created_by_name}</span>}
-            {entry.created_by_name && entry.origin_agent_name && <span style={{ opacity: .5 }}>·</span>}
-            {entry.origin_agent_name && <span>🤖 来源 {entry.origin_agent_name}</span>}
-            {hasCreators && created && <span style={{ opacity: .5 }}>·</span>}
-            {created && <span>创建日期 {created}</span>}
-            {updated && <span>上次修改 {updated}</span>}
-        </div>
-    );
-}
-
 function EntryCard({ entry, onOpen }: { entry: ExperienceEntry; onOpen: () => void }) {
     const f = freshness(entry);
     const daysLeft = retiredDaysLeft(entry);
@@ -419,81 +359,3 @@ function EntryCard({ entry, onOpen }: { entry: ExperienceEntry; onOpen: () => vo
         </div>
     );
 }
-
-function EntryDrawer({ entryId, onClose, onEdit, onChanged }: {
-    entryId: string; onClose: () => void; onEdit: (e: ExperienceEntry) => void; onChanged: () => void;
-}) {
-    const qc = useQueryClient();
-    const { data: entry } = useQuery({ queryKey: ['experience-entry', entryId], queryFn: () => experienceApi.get(entryId) });
-    const { data: refs } = useQuery({ queryKey: ['experience-refs', entryId], queryFn: () => experienceApi.references(entryId) });
-    const retire = useMutation({ mutationFn: () => experienceApi.retire(entryId), onSuccess: () => { onChanged(); onClose(); } });
-    const republish = useMutation({ mutationFn: () => experienceApi.publish(entryId), onSuccess: () => { onChanged(); onClose(); } });
-    const review = useMutation({
-        mutationFn: () => experienceApi.review(entryId),
-        onSuccess: () => { qc.invalidateQueries({ queryKey: ['experience-entry', entryId] }); onChanged(); },
-    });
-
-    if (!entry) return <Drawer onClose={onClose}><div>加载中...</div></Drawer>;
-    const f = freshness(entry);
-    const daysLeft = retiredDaysLeft(entry);
-    return (
-        <Drawer onClose={onClose}>
-            <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'flex-start', gap: 12 }}>
-                <h2 style={{ margin: 0, fontSize: 18, color: 'var(--text-primary)' }}>{entry.title || '(未命名)'}</h2>
-                <button onClick={onClose} style={{ ...secondaryBtn, padding: '4px 10px' }}>✕</button>
-            </div>
-            <div style={{ display: 'flex', gap: 6, flexWrap: 'wrap', margin: '10px 0 16px' }}>
-                <Badge tone="accent">{SCOPE_LABELS[entry.visibility_scope]}</Badge>
-                {f.label && <Badge tone={f.stale ? 'warn' : 'ok'}>{f.label}</Badge>}
-                {(entry.tags || []).map(tg => <Badge key={tg}>#{tg}</Badge>)}
-            </div>
-            <div style={{ marginBottom: 16 }}><CreatorLine entry={entry} /></div>
-            {EXP_FIELDS.map(fl => (
-                <section key={fl.key} style={{ marginBottom: 14 }}>
-                    <div style={{ fontSize: 13, fontWeight: 600, color: 'var(--text-secondary)', marginBottom: 4 }}>{fl.label}</div>
-                    <div style={{ fontSize: 14, color: 'var(--text-primary)', whiteSpace: 'pre-wrap', lineHeight: 1.6 }}>
-                        {(entry[fl.key] as string) || '—'}
-                    </div>
-                </section>
-            ))}
-            {refs && (
-                <div style={{ margin: '8px 0 16px', fontSize: 13, color: 'var(--text-tertiary)' }}>
-                    被 AI 阅读 {refs.read_count} 次 · <strong style={{ color: 'var(--text-secondary)' }}>实际采纳（引用） {refs.cited_count} 次</strong>
-                </div>
-            )}
-            {entry.can_manage ? (
-                <>
-                    <div style={{ display: 'flex', gap: 8, marginTop: 8, flexWrap: 'wrap' }}>
-                        <button onClick={() => onEdit(entry)} style={secondaryBtn}>编辑</button>
-                        {entry.status === 'published' && (
-                            <button onClick={() => review.mutate()} style={secondaryBtn} disabled={review.isPending}>
-                                {entry.last_reviewed_at ? '标记为未复核' : '标记已复核'}
-                            </button>
-                        )}
-                        {entry.status === 'retired' ? (
-                            <button onClick={() => republish.mutate()} style={{ ...secondaryBtn, color: 'var(--success)' }} disabled={republish.isPending}>
-                                重新发布
-                            </button>
-                        ) : (
-                            <button onClick={() => retire.mutate()} style={{ ...secondaryBtn, color: 'var(--error)' }} disabled={retire.isPending}>
-                                下架
-                            </button>
-                        )}
-                    </div>
-                    {entry.status === 'retired' && daysLeft !== null && (
-                        <p style={{ fontSize: 12, color: 'var(--text-tertiary)', marginTop: 10 }}>
-                            已下架，{daysLeft} 天后自动删除。重新发布可恢复到团队经验并清除删除倒计时。
-                        </p>
-                    )}
-                </>
-            ) : (
-                <p style={{ fontSize: 12, color: 'var(--text-tertiary)', marginTop: 10 }}>
-                    {entry.status === 'retired'
-                        ? `已下架${daysLeft !== null ? `，${daysLeft} 天后自动删除` : ''}。仅发起人、数字员工创立者或管理员可编辑或重新发布。`
-                        : '仅发起人、数字员工创立者或管理员可编辑、复核或下架此经验。'}
-                </p>
-            )}
-        </Drawer>
-    );
-}
-
